backend/core/views.py
from .serializers import UserSerializer
from rest_framework import response, viewsets, permissions
from django.contrib.auth.models import User

class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
    permission_classes = (permissions.IsAuthenticated,)